requirements.txt
streamlit
transformers
torch                 # nécessaire pour la majorité des pipelines HF
pytesseract
Pillow
sentencepiece         # parfois requis pour Marian
spacy
PyPDF2